requirements.txt
playwright==1.42.0
beautifulsoup4==4.12.3
pandas==2.2.1
streamlit==1.32.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
schedule
requests
pydantic